serpradio_convo_cache/requirements.txt
openai>=1.35.0
requests>=2.31.0
python-dateutil>=2.9.0.post0
PyYAML>=6.0.1
supabase>=2.5.1
tqdm>=4.66.1
